Finished exercise 3
Finished coding with wes bos for exercise 3 AND finished providing notes
on my research on .sort(). This was a bit difficult for me to understand
only because of how deep the rabbit hole can go in terms of what
circumstances and what things COULD happen while using .sort(). That
being said, I do see how critical it would be for modern day array usage.
I kept the longer function with the if statement as a comment (along
with addition comments for myself) and chose to go with the ternary
operator for a cleaner look.